beef up the quick-loader and swap it in by default
